norm merge

+++ engine/srcs/parser_utility.c
@@ -6,7 +6,7 @@
 /*   By: zytrams <zytrams@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/05 19:12:50 by fsmith            #+#    #+#             */
-/*   Updated: 2019/10/11 19:07:36 by fsmith           ###   ########.fr       */
+/*   Updated: 2019/10/18 19:39:54 by zytrams          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -153,7 +153,7 @@ void		util_create_sector(t_engine *eng, t_buff buff,
 		util_get_object_from_buff_by_id(ft_atoi(str[str_count++]),
 		eng->stats.objects_count, buff.objects, sector->id);
 	util_create_sector_sprobjs(eng, buff, sector, str);
-	sector->dist = (double *)ft_memalloc(sizeof(double) * sector->objects_count);
+	sector->dist = (float *)ft_memalloc(sizeof(double) * sector->objects_count);
 	sector->order = (int *)ft_memalloc(sizeof(int) * sector->objects_count);
 	util_find_repeats_in_sector(sector);
 	eng->stats.sectors_count++;
